feat: rename project to PoCRedis
