+++ src/main/java/com/l10s/PoCRedis/PoCRedisApplication.java
@@ -1,13 +1,13 @@
-package com.l10s.testredis;
+package com.l10s.PoCRedis;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TestredisApplication {
+public class PoCRedisApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(TestredisApplication.class, args);
+        SpringApplication.run(PoCRedisApplication.class, args);
     }
 
 }
