+++ src/main/java/com/l10s/PoCRedis/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.l10s.testredis.redis;
+package com.l10s.PoCRedis.redis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
