+++ src/main/java/com/l10s/PoCRedis/reservations/Reservation.java
@@ -1,8 +1,6 @@
-package com.l10s.testredis.reservations;
+package com.l10s.PoCRedis.reservations;
 import jakarta.persistence.*;
 import lombok.Data;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
