+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 
 spring:
   application:
-    name: testredis
+    name: PoCRedis
   datasource:
     url: jdbc:mysql://localhost:3306/testredis?useSSL=false&serverTimezone=UTC&allowPublicKeyRetrieval=true
     username: appuser
