+++ src/test/java/com/l10s/PoCRedis/PoCRedisApplicationTests.java
@@ -1,10 +1,10 @@
-package com.l10s.testredis;
+package com.l10s.PoCRedis;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TestredisApplicationTests {
+class PoCRedisApplicationTests {
 
     @Test
     void contextLoads() {
